pyproject.toml
[tool.poetry]
name = "calc"
version = "0.1.0"
description = ""
authors = ["Yaroslav Luzin <jardev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
